[PROOFREADING] CSV402 + quiz - French (#4018)
* improvements 1

* second improvements

* third improvements

* fourth improvements

* fifth improvements

* sixth improvements

* Update fr.yml

* Fix wording in quiz question for clarity

* Update fr.yml

* Update fr.yml

* Update fr.yml

* Update fr.yml

* Update fr.yml

* Update fr.yml

* Update fr.yml

Added clarification about business logic in explanation.

* Update fr.yml

Corrected a minor grammatical error in the wrong answer and added clarification to the explanation.

* Update fr.yml

* Update fr.yml

* Update fr.yml

* Update fr.yml

* Update fr.yml

* Update fr.yml

Updated the wording of the answer and wrong answers to use 'financement' instead of 'funding'.

* Fix translation of 'funding' to 'financement'

Corrected the term 'funding' to 'financement' in the wrong answers section.

* Update fr.yml

* Update fr.yml

Corrected the wording for clarity and accuracy in the explanation.

* Update fr.yml

* Update fr.yml

* Update fr.yml

* Fix typos in French quiz questions and answers